pyproject.toml
[project]
name = "oak-curriculum-ontology"
version = "0.1.0"
description = "Oak National Academy Curriculum Ontology - A formal semantic representation of the Oak Curriculum and linkage to the National Curriculum for England (2014)"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "Oak National Academy"}
]
keywords = ["ontology", "curriculum", "education", "rdf", "semantic-web", "shacl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "rdflib>=7.5.0",
    "pyshacl>=0.31.0",
    "rdflib-neo4j>=0.6.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "owlready2>=0.47",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Ignore missing type stubs for external libraries
[[tool.mypy.overrides]]
module = ["rdflib_neo4j", "rdflib_neo4j.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tqdm", "tqdm.*"]
ignore_missing_imports = true

[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
